# backend/docs/swagger.yaml
definitions:
...
  domain.About:
    properties:
      content:
        type: string
      updatedAt:
        type: string
    type: object
  domain.Article:
    properties:
      author:
        type: string
      category:
        type: string
      content:
        type: string
      copyright:
        type: string
      createdAt:
        type: string
      deleted:
        type: boolean
      hidden:
        type: boolean
      id:
        type: integer
      lastVisitedTime:
        type: string
      password:
        type: string
      pathname:
        type: string
      private:
        type: boolean
      tags:
        items:
          type: string
        type: array
      title:
        type: string
      top:
        type: integer
      updatedAt:
        type: string
      viewer:
        type: integer
      visited:
        type: integer
    type: object
  domain.ArticlePageResult:
    properties:
      articles:
        items:
          $ref: '#/definitions/domain.Article'
        type: array
      total:
        type: integer
      totalWordCount:
        type: integer
    type: object
...
  domain.LinkItem:
    properties:
      desc:
        type: string
      logo:
        type: string
      name:
        type: string
      updatedAt:
        type: string
      url:
        type: string
    type: object
...
  domain.MenuItem:
    properties:
      children:
        $ref: '#/definitions/domain.MenuItem'
      id:
        type: integer
      level:
        type: integer
      name:
        type: string
      value:
        type: string
    type: object
  domain.Meta:
    properties:
      about:
        $ref: '#/definitions/domain.About'
      categories:
        items:
          type: string
        type: array
      id:
        type: integer
      links:
        items:
          $ref: '#/definitions/domain.LinkItem'
        type: array
      menus:
        items:
          $ref: '#/definitions/domain.MenuItem'
        type: array
      rewards:
        items:
          $ref: '#/definitions/domain.RewardItem'
        type: array
      siteInfo:
        $ref: '#/definitions/domain.SiteInfo'
      socials:
        items:
          $ref: '#/definitions/domain.SocialItem'
        type: array
      totalWordCount:
        type: integer
      viewer:
        type: integer
      visited:
        type: integer
    type: object
...
  domain.RewardItem:
    properties:
      name:
        type: string
      updatedAt:
        type: string
      value:
        type: string
    type: object
  domain.Site:
    enum:
    - siteLogo
    - siteName
    type: string
    x-enum-varnames:
    - LogoSite
    - NameSite
  domain.SiteInfo:
    properties:
      allowOpenHiddenPostByUrl:
        $ref: '#/definitions/domain.TrueOrFalse'
      authDesc:
        type: string
      author:
        type: string
      authorLogo:
        type: string
      authorLogoDark:
        type: string
      baiduAnalysisId:
        type: string
      baseUrl:
        type: string
      beianNumber:
        type: string
      beianUrl:
        type: string
      copyrightAggreement:
        type: string
      defaultTheme:
        $ref: '#/definitions/domain.Theme'
      enableComment:
        $ref: '#/definitions/domain.TrueOrFalse'
      enableCustomizing:
        $ref: '#/definitions/domain.TrueOrFalse'
      favicon:
        type: string
      gaAnalysisId:
        type: string
      gaBeianLogoUrl:
        type: string
      gaBeianNumber:
        type: string
      gaBeianUrl:
        type: string
      headerLeftContent:
        $ref: '#/definitions/domain.Site'
      openArticleLinksInNewWindow:
        $ref: '#/definitions/domain.TrueOrFalse'
      payAliPay:
        type: string
      payAliPayDark:
        type: string
      payWechat:
        type: string
      payWechatDark:
        type: string
      showAdminButton:
        $ref: '#/definitions/domain.TrueOrFalse'
      showCopyRight:
        $ref: '#/definitions/domain.TrueOrFalse'
      showDonateButton:
        $ref: '#/definitions/domain.TrueOrFalse'
      showDonateInAbout:
        $ref: '#/definitions/domain.TrueOrFalse'
      showDonateInfo:
        $ref: '#/definitions/domain.TrueOrFalse'
      showEditButton:
        $ref: '#/definitions/domain.TrueOrFalse'
      showExpirationReminder:
        $ref: '#/definitions/domain.TrueOrFalse'
      showFriends:
        $ref: '#/definitions/domain.TrueOrFalse'
      showRSS:
        $ref: '#/definitions/domain.TrueOrFalse'
      showSubMenu:
        $ref: '#/definitions/domain.TrueOrFalse'
      since:
        type: string
      siteDesc:
        type: string
      siteLogo:
        type: string
      siteLogoDark:
        type: string
      siteName:
        type: string
      subMenuOffset:
        type: integer
    type: object
  domain.SocialItem:
    properties:
      type:
        $ref: '#/definitions/domain.SocialType'
      updatedAt:
        type: string
      value:
        type: string
    type: object
  domain.SocialType:
    enum:
    - bilibili
    - email
    - github
    - gitee
    - wechat
    - wechat-dark
    type: string
    x-enum-varnames:
    - BilibiliSocialType
    - EmailSocialType
    - GithubSocialType
    - GiteeSocialType
    - WechatSocialType
    - WechatDarkSocialType
...
  domain.Theme:
    enum:
    - auto
    - dark
    - light
    type: string
    x-enum-varnames:
    - AutoTheme
    - DarkTheme
    - LightTheme
...
  domain.TrueOrFalse:
    enum:
    - "true"
    - "false"
    type: string
    x-enum-varnames:
    - TrueString
    - FalseString
  domain.User:
    properties:
      createdAt:
        type: string
      id:
        type: integer
      name:
        type: string
      nickname:
        type: string
      password:
        type: string
      permissions:
        items:
          type: string
        type: array
      salt:
        type: string
      type:
        type: string
    type: object
...
  svr.TokenUser:
    properties:
      token:
        type: string
      user:
        $ref: '#/definitions/domain.User'
    type: object
info:
  contact: {}
paths:
  /api/admin/analysis:
    get:
      consumes:
      - application/json
      description: statistic home page data
      parameters:
      - description: tab
        enum:
        - overview
        - viewer
        - article
        in: query
        name: tab
        required: true
        type: string
      - description: overviewDataNum
        in: query
        name: viewerDataNum
        type: integer
      - description: overviewDataNum
        in: query
        name: overviewDataNum
        type: integer
      - description: articleTabDataNum
        in: query
        name: articleTabDataNum
        type: integer
      produces:
      - application/json
      responses: {}
      summary: 获取首页统计数据
      tags:
      - Analysis
  /api/admin/article:
    get:
      consumes:
      - application/json
      description: from option obtain articles
      parameters:
      - description: page
        in: query
        name: page
        type: integer
      - description: pageSize
        in: query
        name: pageSize
        type: integer
      - description: toListView
        in: query
        name: toListView
        type: boolean
      - description: category
        in: query
        name: category
        type: string
      - description: page
        in: query
        name: tags
        type: string
      - description: title
        in: query
        name: title
        type: integer
      - description: sortCreateAt
        in: query
        name: sortCreateAt
        type: integer
      - description: sortTop
        in: query
        name: sortTop
        type: integer
      - description: sortViewer
        in: query
        name: sortViewer
        type: integer
      - description: startTime
        in: query
        name: startTime
        type: integer
      - description: endTime
        in: query
        name: endTime
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.ArticlePageResult'
      summary: 根据参数获取文章
      tags:
      - Article
  /api/admin/article/:id:
    delete:
      consumes:
      - application/json
      description: 删除文章
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: boolean
      summary: 删除文章
      tags:
      - Article
    get:
      consumes:
      - application/json
      description: 根据id获取article
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.ArticlePageResult'
      summary: 根据id获取article
      tags:
      - Article
    post:
      consumes:
      - application/json
      description: 创建文章
      parameters:
      - description: query params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.Article'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: boolean
      summary: 创建文章
      tags:
      - Article
    put:
      consumes:
      - application/json
      description: 根据id获取article
      parameters:
      - description: query params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.Article'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: boolean
      summary: 根据id更新文档
      tags:
      - Article
...
  /api/admin/auth/login:
    post:
      consumes:
      - application/json
      description: login by username and password
      parameters:
      - description: username
        in: path
        name: username
        required: true
        type: string
      - description: password
        in: path
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/svr.TokenUser'
      summary: 登陆
      tags:
      - Auth
...
  /api/admin/backup/import:
    post:
      consumes:
      - application/json
      description: 导入系统数据
      parameters:
      - description: file
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: 导入系统数据
      tags:
      - Backup
...
  /api/admin/meta/about:
    get:
      consumes:
      - application/json
      description: obtain about information
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Meta'
      summary: 获取About
      tags:
      - Meta
    put:
      consumes:
      - application/json
      description: obtain about information
      parameters:
      - description: content
        in: path
        name: content
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: boolean
      summary: 更新 about
      tags:
      - Meta
...
swagger: "2.0"